chore: rename "serie" -> "series"
