+++ tests/series/mod.rs
@@ -1,2 +1,2 @@
-mod serie_browse;
+mod series_browse;
 mod series_includes;
