+++ tests/series/series_browse.rs
