+++ tests/series/series_includes.rs
@@ -3,7 +3,7 @@ use musicbrainz_rs::entity::series::Series;
 use musicbrainz_rs::prelude::*;
 
 #[test]
-fn should_get_serie_tags() {
+fn should_get_series_tags() {
     let breaks_loop_n_edits = Series::fetch()
         .id("0c66e70d-5f23-4579-8fe5-6bc0007428a2")
         .with_tags()
@@ -18,7 +18,7 @@ fn should_get_serie_tags() {
 }
 
 #[test]
-fn should_get_serie_aliases() {
+fn should_get_series_aliases() {
     let ultimate_breaks_and_beats = Series::fetch()
         .id("3e5979c8-5a78-4d0b-878a-0fb87853effe")
         .with_aliases()
@@ -29,7 +29,7 @@ fn should_get_serie_aliases() {
 }
 
 #[test]
-fn should_get_serie_genres() {
+fn should_get_series_genres() {
     let ultimate_breaks_and_beats = Series::fetch()
         .id("3e5979c8-5a78-4d0b-878a-0fb87853effe")
         .with_genres()
@@ -40,7 +40,7 @@ fn should_get_serie_genres() {
 }
 
 #[test]
-fn should_get_serie_annotation() {
+fn should_get_series_annotation() {
     let record_store_day_2020 = Series::fetch()
         .id("c1071cec-48f1-4231-ac8e-8c64e15ec7cd")
         .with_annotations()
